refactor okhttp to own jar
