+++ jaxrs-clients/build.gradle
@@ -7,6 +7,7 @@ dependencies {
     compile project(":error-handling")
     compile project(":tracing-okhttp3")
     compile project(":http-clients-api")
+    compile project(":okhttp-clients")
     compile project(":service-config")
 
     compile "com.google.code.findbugs:jsr305"
