+++ okhttp-clients/build.gradle
@@ -1,0 +1,9 @@
+apply from: "${rootDir}/gradle/jaxrs-shading-config.gradle"
+apply from: "${rootDir}/gradle/publish-shadow.gradle"
+
+dependencies {
+    compile project(":tracing-okhttp3")
+    compile project(":http-clients-api")
+
+    compile "org.slf4j:slf4j-api"
+}
