+++ okhttp-clients/src/main/java/com/palantir/remoting2/okhttp/RetryInterceptor.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.palantir.remoting2.retrofit2;
+package com.palantir.remoting2.okhttp;
 
 import java.io.IOException;
 import okhttp3.Interceptor;
