+++ retrofit2-clients/build.gradle
@@ -8,6 +8,7 @@ dependencies {
     compile project(":extras:refresh-utils")
     compile project(":error-handling")
     compile project(":http-clients-api")
+    compile project(":okhttp-clients")
     compile project(":service-config")
     compile project(":tracing-okhttp3")
     // ssl-config is not used in this project, but we depend on it here in order to provide
