+++ retrofit2-clients/src/main/java/com/palantir/remoting2/retrofit2/Retrofit2ClientBuilder.java
@@ -18,25 +18,15 @@ package com.palantir.remoting2.retrofit2;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.base.Preconditions;
-import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Lists;
-import com.google.common.net.HttpHeaders;
-import com.palantir.remoting2.clients.CipherSuites;
 import com.palantir.remoting2.clients.ClientBuilder;
 import com.palantir.remoting2.clients.ClientConfig;
-import com.palantir.remoting2.config.service.BasicCredentials;
-import com.palantir.remoting2.config.service.ProxyConfiguration;
-import com.palantir.remoting2.config.ssl.TrustContext;
 import com.palantir.remoting2.ext.jackson.ObjectMappers;
-import com.palantir.remoting2.tracing.okhttp3.OkhttpTraceInterceptor;
+import com.palantir.remoting2.okhttp.MultiServerRetryInterceptor;
+import com.palantir.remoting2.okhttp.OkHttpClients;
 import java.util.List;
-import java.util.concurrent.TimeUnit;
-import okhttp3.ConnectionPool;
-import okhttp3.ConnectionSpec;
-import okhttp3.Credentials;
 import okhttp3.Dispatcher;
 import okhttp3.OkHttpClient;
-import okhttp3.TlsVersion;
 import retrofit2.Retrofit;
 import retrofit2.converter.jackson.JacksonConverterFactory;
 
@@ -78,50 +68,11 @@ public final class Retrofit2ClientBuilder extends ClientBuilder {
     }
 
     private OkHttpClient createOkHttpClient(String userAgent, List<String> uris) {
-        OkHttpClient.Builder client = new OkHttpClient.Builder();
-
-        // SSL
-        if (config.trustContext().isPresent()) {
-            TrustContext context = config.trustContext().get();
-            client.sslSocketFactory(context.sslSocketFactory(), context.x509TrustManager());
-        }
-
-        // proxy
-        if (config.proxy().isPresent()) {
-            ProxyConfiguration proxy = config.proxy().get();
-            client.proxy(proxy.toProxy());
-
-            if (proxy.credentials().isPresent()) {
-                BasicCredentials proxyCredentials = proxy.credentials().get();
-                final String credentials = Credentials.basic(proxyCredentials.username(), proxyCredentials.password());
-                client.proxyAuthenticator((route, response) -> response.request().newBuilder()
-                        .header(HttpHeaders.PROXY_AUTHORIZATION, credentials)
-                        .build());
-            }
-        }
-
-        // tracing
-        client.interceptors().add(OkhttpTraceInterceptor.INSTANCE);
-
-        // timeouts
-        client.connectTimeout(config.connectTimeout().toMilliseconds(), TimeUnit.MILLISECONDS);
-        client.readTimeout(config.readTimeout().toMilliseconds(), TimeUnit.MILLISECONDS);
-        client.writeTimeout(config.writeTimeout().toMilliseconds(), TimeUnit.MILLISECONDS);
-
-        // retry configuration
-        if (config.maxNumRetries() > 1) {
-            client.addInterceptor(new RetryInterceptor(config.maxNumRetries()));
-        }
+        OkHttpClient.Builder client = OkHttpClients.createClient(config).newBuilder();
 
         client.addInterceptor(MultiServerRetryInterceptor.create(uris));
         client.addInterceptor(UserAgentInterceptor.of(userAgent));
         client.addInterceptor(SerializableErrorInterceptor.INSTANCE);
-
-        // cipher setup
-        client.connectionSpecs(createConnectionSpecs(config.enableGcmCipherSuites()));
-
-        // increase default connection pool from 5 @ 5 minutes to 100 @ 10 minutes
-        client.connectionPool(new ConnectionPool(100, 10, TimeUnit.MINUTES));
 
         Dispatcher dispatcher = new Dispatcher();
 
@@ -133,14 +84,4 @@ public final class Retrofit2ClientBuilder extends ClientBuilder {
         return client.build();
     }
 
-    private static ImmutableList<ConnectionSpec> createConnectionSpecs(boolean enableGcmCipherSuites) {
-        return ImmutableList.of(
-            new ConnectionSpec.Builder(ConnectionSpec.MODERN_TLS)
-                    .tlsVersions(TlsVersion.TLS_1_2)
-                    .cipherSuites(enableGcmCipherSuites
-                            ? CipherSuites.allCipherSuites()
-                            : CipherSuites.fastCipherSuites())
-                    .build(),
-            ConnectionSpec.CLEARTEXT);
-    }
 }
