+++ settings.gradle
@@ -4,6 +4,7 @@ include 'error-handling'
 include 'http-clients-api'
 include 'http2-agent'
 include 'jaxrs-clients'
+include 'okhttp-clients'
 include 'jaxrs-scala-clients'
 include 'jaxrs-clients-compat-test'
 include 'jersey-servers'
